# Conflicts:
#	package.json
#	src/legacy/core_plugins/kibana/public/dashboard/np_ready/listing/__snapshots__/dashboard_listing.test.js.snap
#	src/legacy/core_plugins/kibana/public/management/sections/index_patterns/edit_index_pattern/scripted_fields_table/components/call_outs/__jest__/__snapshots__/call_outs.test.js.snap
#	src/legacy/core_plugins/kibana/public/management/sections/objects/components/objects_table/components/flyout/__jest__/__snapshots__/flyout.test.js.snap
#	x-pack/legacy/plugins/monitoring/public/components/metricbeat_migration/flyout/__snapshots__/flyout.test.js.snap
#	x-pack/legacy/plugins/siem/public/components/page/add_filter_to_global_search_bar/__snapshots__/index.test.tsx.snap
#	x-pack/legacy/plugins/uptime/public/components/functional/empty_state/__tests__/__snapshots__/data_missing.test.tsx.snap
#	x-pack/legacy/plugins/uptime/public/components/functional/empty_state/__tests__/__snapshots__/empty_state.test.tsx.snap
#	x-pack/legacy/plugins/uptime/public/components/functional/ping_list/__tests__/__snapshots__/ping_list.test.tsx.snap
#	yarn.lock

+++ x-pack/plugins/actions/server/actions_client.ts
@@ -11,9 +11,16 @@ import {
   SavedObject,
 } from 'src/core/server';
 
+import { i18n } from '@kbn/i18n';
 import { ActionTypeRegistry } from './action_type_registry';
 import { validateConfig, validateSecrets } from './lib';
-import { ActionResult, FindActionResult, RawAction } from './types';
+import { ActionResult, FindActionResult, RawAction, PreConfiguredAction } from './types';
+import { PreconfiguredActionDisabledModificationError } from './lib/errors/preconfigured_action_disabled_modification';
+
+// We are assuming there won't be many actions. This is why we will load
+// all the actions in advance and assume the total count to not go over 10000.
+// We'll set this max setting assuming it's never reached.
+export const MAX_ACTIONS_RETURNED = 10000;
 
 interface ActionUpdate extends SavedObjectAttributes {
   name: string;
@@ -29,35 +36,12 @@ interface CreateOptions {
   action: Action;
 }
 
-interface FindOptions {
-  options?: {
-    perPage?: number;
-    page?: number;
-    search?: string;
-    defaultSearchOperator?: 'AND' | 'OR';
-    searchFields?: string[];
-    sortField?: string;
-    hasReference?: {
-      type: string;
-      id: string;
-    };
-    fields?: string[];
-    filter?: string;
-  };
-}
-
-interface FindResult {
-  page: number;
-  perPage: number;
-  total: number;
-  data: FindActionResult[];
-}
-
 interface ConstructorOptions {
   defaultKibanaIndex: string;
   scopedClusterClient: IScopedClusterClient;
   actionTypeRegistry: ActionTypeRegistry;
   savedObjectsClient: SavedObjectsClientContract;
+  preconfiguredActions: PreConfiguredAction[];
 }
 
 interface UpdateOptions {
@@ -70,17 +54,20 @@ export class ActionsClient {
   private readonly scopedClusterClient: IScopedClusterClient;
   private readonly savedObjectsClient: SavedObjectsClientContract;
   private readonly actionTypeRegistry: ActionTypeRegistry;
+  private readonly preconfiguredActions: PreConfiguredAction[];
 
   constructor({
     actionTypeRegistry,
     defaultKibanaIndex,
     scopedClusterClient,
     savedObjectsClient,
+    preconfiguredActions,
   }: ConstructorOptions) {
     this.actionTypeRegistry = actionTypeRegistry;
     this.savedObjectsClient = savedObjectsClient;
     this.scopedClusterClient = scopedClusterClient;
     this.defaultKibanaIndex = defaultKibanaIndex;
+    this.preconfiguredActions = preconfiguredActions;
   }
 
   /**
@@ -106,6 +93,7 @@ export class ActionsClient {
       actionTypeId: result.attributes.actionTypeId,
       name: result.attributes.name,
       config: result.attributes.config,
+      isPreconfigured: false,
     };
   }
 
@@ -113,6 +101,20 @@ export class ActionsClient {
    * Update action
    */
   public async update({ id, action }: UpdateOptions): Promise<ActionResult> {
+    if (
+      this.preconfiguredActions.find(preconfiguredAction => preconfiguredAction.id === id) !==
+      undefined
+    ) {
+      throw new PreconfiguredActionDisabledModificationError(
+        i18n.translate('xpack.actions.serverSideErrors.predefinedActionUpdateDisabled', {
+          defaultMessage: 'Preconfigured action {id} is not allowed to update.',
+          values: {
+            id,
+          },
+        }),
+        'update'
+      );
+    }
     const existingObject = await this.savedObjectsClient.get<RawAction>('action', id);
     const { actionTypeId } = existingObject.attributes;
     const { name, config, secrets } = action;
@@ -133,7 +135,8 @@ export class ActionsClient {
       id,
       actionTypeId: result.attributes.actionTypeId as string,
       name: result.attributes.name as string,
-      config: result.attributes.config as Record<string, any>,
+      config: result.attributes.config as Record<string, unknown>,
+      isPreconfigured: false,
     };
   }
 
@@ -141,6 +144,17 @@ export class ActionsClient {
    * Get an action
    */
   public async get({ id }: { id: string }): Promise<ActionResult> {
+    const preconfiguredActionsList = this.preconfiguredActions.find(
+      preconfiguredAction => preconfiguredAction.id === id
+    );
+    if (preconfiguredActionsList !== undefined) {
+      return {
+        id,
+        actionTypeId: preconfiguredActionsList.actionTypeId,
+        name: preconfiguredActionsList.name,
+        isPreconfigured: true,
+      };
+    }
     const result = await this.savedObjectsClient.get<RawAction>('action', id);
 
     return {
@@ -148,36 +162,55 @@ export class ActionsClient {
       actionTypeId: result.attributes.actionTypeId,
       name: result.attributes.name,
       config: result.attributes.config,
+      isPreconfigured: false,
     };
   }
 
   /**
-   * Find actions
+   * Get all actions with preconfigured list
    */
-  public async find({ options = {} }: FindOptions): Promise<FindResult> {
-    const findResult = await this.savedObjectsClient.find<RawAction>({
-      ...options,
-      type: 'action',
-    });
+  public async getAll(): Promise<FindActionResult[]> {
+    const savedObjectsActions = (
+      await this.savedObjectsClient.find<RawAction>({
+        perPage: MAX_ACTIONS_RETURNED,
+        type: 'action',
+      })
+    ).saved_objects.map(actionFromSavedObject);
 
-    const data = await injectExtraFindData(
+    const mergedResult = [
+      ...savedObjectsActions,
+      ...this.preconfiguredActions.map(preconfiguredAction => ({
+        id: preconfiguredAction.id,
+        actionTypeId: preconfiguredAction.actionTypeId,
+        name: preconfiguredAction.name,
+        isPreconfigured: true,
+      })),
+    ].sort((a, b) => a.name.localeCompare(b.name));
+    return await injectExtraFindData(
       this.defaultKibanaIndex,
       this.scopedClusterClient,
-      findResult.saved_objects.map(actionFromSavedObject)
+      mergedResult
     );
-
-    return {
-      page: findResult.page,
-      perPage: findResult.per_page,
-      total: findResult.total,
-      data,
-    };
   }
 
   /**
    * Delete action
    */
   public async delete({ id }: { id: string }) {
+    if (
+      this.preconfiguredActions.find(preconfiguredAction => preconfiguredAction.id === id) !==
+      undefined
+    ) {
+      throw new PreconfiguredActionDisabledModificationError(
+        i18n.translate('xpack.actions.serverSideErrors.predefinedActionDeleteDisabled', {
+          defaultMessage: 'Preconfigured action {id} is not allowed to delete.',
+          values: {
+            id,
+          },
+        }),
+        'delete'
+      );
+    }
     return await this.savedObjectsClient.delete('action', id);
   }
 }
@@ -186,6 +219,7 @@ function actionFromSavedObject(savedObject: SavedObject<RawAction>): ActionResul
   return {
     id: savedObject.id,
     ...savedObject.attributes,
+    isPreconfigured: false,
   };
 }
 
@@ -194,7 +228,7 @@ async function injectExtraFindData(
   scopedClusterClient: IScopedClusterClient,
   actionResults: ActionResult[]
 ): Promise<FindActionResult[]> {
-  const aggs: Record<string, any> = {};
+  const aggs: Record<string, unknown> = {};
   for (const actionResult of actionResults) {
     aggs[actionResult.id] = {
       filter: {
